tbio: fix cmd_len handling
sizeof(rq->cmd) does not return the size of cmd buffer in rq,
it will return the size of the pointer, i.e. 4 - on 32-bit systems,
and 8 - on 64-bit systems. This happens, because cmd is a pointer to
__cmd[BLK_MAX_CDB] in struct request in the kernel (see linux/blkdev.h):

        unsigned char __cmd[BLK_MAX_CDB];
        unsigned char *cmd;
        unsigned short cmd_len;

and is not a static buffer by itself.

Therefore on 32-bit systems 'sizeof(rq->cmd) - inter->cmd_len)' will be
'4 - 6', and this is not what we want.

Corrected this, so now we rely on inter->cmd containing  a null-terminated string,
and inter->cmd_len containing the exact number of bytes to store this string.

Acked-by: Cyril Hrubis <chrubis@suse.cz>

+++ testcases/kernel/device-drivers/tbio/tbio_kernel/tbio.h
@@ -66,6 +66,6 @@ struct tbio_interface {
 	int data_len;	/* input data length */
 	int direction;	/* read or write form DEV */
 	char *cmd;	/* read or write */
-	int cmd_len;	/* length of cmd */
+	unsigned short cmd_len;	/* length of cmd */
 };
 typedef struct tbio_interface tbio_interface_t;
